chore: move the pages up a level
Since we're using `pathPrefix` we no longer have to place the pages themselves
in a `docs` directory, everything can be pulled up to the root of the pages
folder
