+++ src/pages/products/livesync.tsx
@@ -1,11 +1,11 @@
-import { graphql } from 'gatsby';
 import { Helmet } from 'react-helmet';
+import { graphql } from 'gatsby';
 
 import Layout from 'src/components/Layout';
+import { LeftSideBar } from 'src/components/StaticQuerySidebar';
 import { ProductPageContent, SectionProps } from 'src/components/ProductPage/ProductPageContent';
 
-import { SidebarProvider } from 'src/contexts/SidebarContext';
-import { DOCUMENTATION_NAME } from '../../../../data/transform/constants';
+import { DOCUMENTATION_NAME } from '../../../data/transform/constants';
 
 type MetaData = {
   title: string;
@@ -32,8 +32,8 @@ const IndexPage = ({
         <meta property="og:title" content={openGraphTitle} />
         <meta property="twitter:title" content={openGraphTitle} />
         <meta property="og:site_name" content="Ably Realtime" />
-        <link rel="canonical" href={`/${DOCUMENTATION_NAME}/products/spaces`} />
-        <meta property="og:url" content={`${ABLY_MAIN_WEBSITE}/${DOCUMENTATION_NAME}/products/spaces`} />
+        <link rel="canonical" href={`/${DOCUMENTATION_NAME}/products/livesync`} />
+        <meta property="og:url" content={`${ABLY_MAIN_WEBSITE}/${DOCUMENTATION_NAME}/products/livesync`} />
         <meta name="description" content={meta.description} />
         <meta property="og:description" content={meta.description} />
         <meta name="twitter:description" content={meta.description} />
@@ -42,18 +42,17 @@ const IndexPage = ({
         <meta name="twitter:image" content={meta.image} />
       </Helmet>
 
-      <SidebarProvider>
-        <Layout isExtraWide currentProduct="spaces">
-          <ProductPageContent sections={sections} />
-        </Layout>
-      </SidebarProvider>
+      <Layout isExtraWide currentProduct="livesync">
+        <LeftSideBar sidebarName="livesync" />
+        <ProductPageContent sections={sections} />
+      </Layout>
     </>
   );
 };
 
 export const query = graphql`
   query HomePageQuery {
-    pageContentYaml(name: { eq: "Spaces" }) {
+    pageContentYaml(name: { eq: "LiveSync" }) {
       sections {
         title
         level
@@ -71,6 +70,7 @@ export const query = graphql`
           type
           content
           image
+          link
           links {
             text
             href
